+++ library/src/main/java/com/davidstemmer/screenplay/flow/Screenplay.java
@@ -14,73 +14,94 @@ import java.util.Iterator;
 import flow.Backstack;
 import flow.Flow;
 
+
+/**
+ * The Screenplay object handles the navigation logic for a Screenplay application.
+ */
 public class Screenplay implements Flow.Listener {
 
     private SceneState screenState = SceneState.NORMAL;
 
     private final Stage stage;
 
-    private final ArrayDeque<Scene> previousScenes = new ArrayDeque<Scene>();
+    private final ArrayDeque<Scene> previousBackstack = new ArrayDeque<>();
 
     public Screenplay(Stage stage) {
         this.stage = stage;
     }
 
     @Override
-    public void go(Backstack nextBackstack, Flow.Direction direction, Flow.Callback callback) {
+    public void go(Backstack backstack, Flow.Direction direction, Flow.Callback callback) {
 
-        Scene incomingScene = (Scene) nextBackstack.current().getScreen();
+        ArrayDeque<Scene> nextBackstack = deque(backstack);
 
         ArrayDeque<Scene> incomingScenes;
-        ArrayDeque<Scene> outgoingScenes;
+        ArrayDeque<Scene> outoingScenes;
 
         SceneCut.Builder sceneCut = new SceneCut.Builder()
                 .setDirection(direction)
                 .setCallback(callback);
 
         if (direction == Flow.Direction.BACKWARD) {
-            Scene outgoingScene = previousScenes.getLast();
-            if (outgoingScene.isStacking()) {
-                incomingScenes = new ArrayDeque<>();
-            } else {
-                incomingScenes = getLastSceneBlock(nextBackstack);
-            }
-            outgoingScenes = new ArrayDeque<>();
-            outgoingScenes.push(outgoingScene);
+            ArrayDeque<Scene> difference = difference(previousBackstack, nextBackstack);
+            outoingScenes = getLastSceneBlock(difference);
+            incomingScenes = moveToNewSceneBlock(difference) ?
+                    getLastSceneBlock(nextBackstack) :
+                    new ArrayDeque<Scene>();
         }
         else {
-            // Using forward/replace, there is always one (and only one) incoming scene
-            incomingScenes = new ArrayDeque<>();
-            incomingScenes.push(incomingScene);
-            // Forward/replace to a stacked scene:
-            // No outgoing scenes to animate or tear down
-            if (incomingScene.isStacking() || previousScenes.isEmpty()) {
-                outgoingScenes = new ArrayDeque<>();
-            }
-            // Forward/replace to a non-stacked scene:
-            // Animate and tear down the previous scene block
-            else {
-                outgoingScenes = previousScenes;
-            }
+            ArrayDeque<Scene> difference = difference(nextBackstack, previousBackstack);
+
+            incomingScenes = getLastSceneBlock(difference);
+            outoingScenes = moveToNewSceneBlock(difference) ?
+                    getLastSceneBlock(previousBackstack) :
+                    new ArrayDeque<Scene>();
 
         }
 
         Scene.Transformer delegatedTransformer = direction == Flow.Direction.BACKWARD ?
-                outgoingScenes.getFirst().getTransformer():
-                incomingScene.getTransformer();
+                outoingScenes.iterator().next().getTransformer():
+                incomingScenes.iterator().next().getTransformer();
 
-        for (Scene scene : incomingScenes) {
-            sceneCut.addIncomingScene(scene);
-        }
+        sceneCut.setIncomingScenes(incomingScenes);
+        sceneCut.setOutgoingScenes(outoingScenes);
 
-        for (Scene scene : outgoingScenes) {
-            sceneCut.addOutgoingScene(scene);
-        }
-
-        previousScenes.clear();
-        previousScenes.addAll(getLastSceneBlock(nextBackstack));
+        previousBackstack.clear();
+        previousBackstack.addAll(nextBackstack);
 
         beginCut(sceneCut.build(), delegatedTransformer);
+    }
+
+    private ArrayDeque<Scene> deque(Backstack backstack) {
+        ArrayDeque<Scene> sceneSet = new ArrayDeque<Scene>();
+        for (Backstack.Entry entry : backstack) {
+            Scene nextScene = (Scene) entry.getScreen();
+            sceneSet.add(nextScene);
+        }
+        return sceneSet;
+    }
+
+    private boolean moveToNewSceneBlock(ArrayDeque<Scene> difference) {
+        for (Scene scene : difference) {
+            if (!scene.isStacking()) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    private ArrayDeque<Scene> difference(ArrayDeque<Scene> sourceSet,
+                                         ArrayDeque<Scene> subSet) {
+        ArrayDeque<Scene> difference = new ArrayDeque<>(sourceSet);
+        difference.removeAll(subSet);
+        return validateDifference(difference);
+    }
+
+    private ArrayDeque<Scene> validateDifference(ArrayDeque<Scene> difference) {
+        if (difference.size() == 0) {
+            throw new IllegalStateException("Backstack validation error -- is the same Scene instance being added to the backstack more than once?");
+        }
+        return difference;
     }
 
     /**
@@ -110,30 +131,30 @@ public class Screenplay implements Flow.Listener {
         return false;
     }
 
-    private Backstack trimBackstack(Backstack backstack, int numItems) {
-        Backstack.Builder builder = backstack.buildUpon();
-        for (int i = 0; i< numItems; i++) {
-            builder.pop();
-        }
-        return builder.build();
-    }
-
-    private static ArrayDeque<Scene> getLastSceneBlock(Backstack backstack) {
-        Iterator<Backstack.Entry> iterator = backstack.iterator();
-        ArrayDeque<Scene> stackedScenes = new ArrayDeque<>();
-
-        while(iterator.hasNext()) {
-            Scene scene = (Scene) iterator.next().getScreen();
-            stackedScenes.push(scene);
+    /**
+     * A "scene block" is a group of associated scenes. A scene block is derived from a source list
+     * by iterating (starting from the first/topmost scene) until a non-stacking scene is found
+     * or the source list terminates.
+     * @param difference the difference between the old backstack and the new backstack
+     * @return a scene block to animate
+     */
+    private ArrayDeque<Scene> getLastSceneBlock(ArrayDeque<Scene> difference) {
+        ArrayDeque<Scene> sceneBlock = new ArrayDeque<>();
+        Iterator<Scene> sceneIterator = difference.iterator();
+        while(sceneIterator.hasNext()) {
+            Scene scene = sceneIterator.next();
+            sceneBlock.add(scene);
             if (!scene.isStacking()) {
-                return stackedScenes;
+                break;
             }
         }
-        throw new IllegalStateException("Modal view cannot be the root of a scene stack.");
+        return sceneBlock;
     }
 
     public void beginCut(SceneCut cut, Scene.Transformer transformer) {
-        for (Scene scene : cut.incomingScenes) {
+        Iterator<Scene> decendingIterator = cut.incomingScenes.descendingIterator();
+        while (decendingIterator.hasNext()) {
+            Scene scene = decendingIterator.next();
             boolean isStarting = cut.direction != Flow.Direction.BACKWARD;
             setUpScene(scene, isStarting);
             setUpComponents(scene, isStarting);
@@ -148,7 +169,9 @@ public class Screenplay implements Flow.Listener {
      * @param cut contains the next and previous scene, and the flow direction
      */
     public void endCut(SceneCut cut) {
-        for (Scene scene : cut.outgoingScenes) {
+        Iterator<Scene> iterator = cut.outgoingScenes.iterator();
+        while (iterator.hasNext()) {
+            Scene scene = iterator.next();
             boolean isFinishing = cut.direction != Flow.Direction.FORWARD;
             tearDownComponents(scene, isFinishing);
             tearDownScene(scene, isFinishing);
@@ -175,33 +198,33 @@ public class Screenplay implements Flow.Listener {
             throw new IllegalStateException("Backstack is empty");
         }
 
-        ArrayDeque<Scene> scenes = getLastSceneBlock(flow.getBackstack());
-        Iterator<Scene> sceneIterator = scenes.iterator();
+        ArrayDeque<Scene> scenes = getLastSceneBlock(deque(flow.getBackstack()));
 
-        Scene firstScene = scenes.getFirst();
-        if (scenes.size() == 1 && previousScenes.isEmpty()) {
-            flow.replaceTo(firstScene);
-        }
-        else {
-            while(sceneIterator.hasNext()) {
-                Scene nextScene = sceneIterator.next();
-                if (nextScene.teardownOnConfigurationChange()) {
-                    setUpScene(nextScene, false);
-                    setUpComponents(nextScene, false);
-                } else {
-                    attachToParent(stage, nextScene.getView());
-                }
+        // the topmost scene is the first in the scene block
+        // re-add the scenes in reverse order, from bottom top
+        Iterator<Scene> sceneIterator = scenes.descendingIterator();
+
+        while(sceneIterator.hasNext()) {
+            Scene nextScene = sceneIterator.next();
+            if (nextScene.teardownOnConfigurationChange() || previousBackstack.isEmpty()) {
+                setUpScene(nextScene, false);
+                setUpComponents(nextScene, false);
+            } else {
+                attachToParent(stage, nextScene.getView());
             }
         }
     }
 
     public void exit() {
 
-        ArrayDeque<Scene> outgoingScenes = getLastSceneBlock(stage.getFlow().getBackstack());
-        Iterator<Scene> reverseSceneBlockIterator = outgoingScenes.descendingIterator();
+        ArrayDeque<Scene> outgoingScenes = new ArrayDeque<>(getLastSceneBlock(previousBackstack));
 
-        while (reverseSceneBlockIterator.hasNext()) {
-            Scene nextScene = reverseSceneBlockIterator.next();
+        // the topmost scene is the first in the scene block
+        // tear down the scenes in order, from top to bottom
+        Iterator<Scene> sceneBlockIterator = outgoingScenes.iterator();
+
+        while(sceneBlockIterator.hasNext()) {
+            Scene nextScene = sceneBlockIterator.next();
             if (nextScene.teardownOnConfigurationChange()) {
                 tearDownComponents(nextScene, false);
                 tearDownScene(nextScene, false);
@@ -215,9 +238,20 @@ public class Screenplay implements Flow.Listener {
         stage.getContainer().addView(view);
     }
 
-    private void removeFromParent(View view) {
-        ViewGroup parent = (ViewGroup) view.getParent();
-        parent.removeView(view);
+    private void removeFromParent(final View view) {
+
+        // If we don't include the post() call, the following error occurs:
+        // Attempt to read from field 'int android.view.View.mViewFlags' on a null object reference
+        // So far I haven't been able to isolate the cause of this -- I think it might be related
+        // to removing the view immediately after an animation ends, but I haven't been able to
+        // reproduce it under controlled conditions.
+        final ViewGroup parent = (ViewGroup) view.getParent();
+        parent.post(new Runnable() {
+            @Override
+            public void run() {
+                parent.removeView(view);
+            }
+        });
     }
 
     private void setUpScene(Scene scene, boolean isStarting) {
@@ -239,7 +273,6 @@ public class Screenplay implements Flow.Listener {
     private void tearDownComponents(Scene scene, boolean isFinishing) {
         for (Scene.Component component: scene.getComponents()) {
             component.beforeTearDown(scene, isFinishing);
-
         }
     }
 }