+++ sample-simple/src/main/java/com/davidstemmer/screenplay/sample/simple/scene/StaticScenes.java
@@ -1,0 +1,20 @@
+package com.davidstemmer.screenplay.sample.simple.scene;
+
+/**
+ * Created by weefbellington on 6/18/15.
+ */
+public class StaticScenes {
+
+    public static final WelcomeScene WELCOME_SCENE = new WelcomeScene();
+    public static final PagedScene1 SIMPLE_PAGED_SCENE_1 = new PagedScene1();
+    public static final PagedScene2 SIMPLE_PAGED_SCENE_2 = new PagedScene2();
+    public static final PagedScene3 SIMPLE_PAGED_SCENE_3 = new PagedScene3();
+    public static final ComplexPagedScene1 COMPLEX_PAGED_SCENE_1 = new ComplexPagedScene1();
+    public static final ComplexPagedScene2 COMPLEX_PAGED_SCENE_2 = new ComplexPagedScene2();
+    public static final ComplexPagedScene3 COMPLEX_PAGED_SCENE_3 = new ComplexPagedScene3();
+    public static final NewBackstackScene1 NEW_BACKSTACK_SCENE_1 = new NewBackstackScene1();
+    public static final NewBackstackScene2 NEW_BACKSTACK_SCENE_2 = new NewBackstackScene2();
+    public static final ModalViewsScene MODAL_VIEWS_SCENE = new ModalViewsScene();
+    public static final DialogScene DIALOG_SCENE = new DialogScene();
+    public static final PostDialogScene POST_DIALOG_SCENE = new PostDialogScene();
+}