+++ sample-simple/src/main/java/com/davidstemmer/screenplay/sample/simple/scene/WelcomeScene2.java
@@ -1,5 +1,7 @@
 package com.davidstemmer.screenplay.sample.simple.scene;
 
+import android.app.Application;
+
 import com.davidstemmer.screenplay.sample.simple.SampleApplication;
 import com.davidstemmer.screenplay.sample.simple.scene.transformer.CrossfadeTransformer;
 import com.example.weefbellington.screenplay.sample.simple.R;
@@ -9,14 +11,19 @@ import flow.Layout;
 /**
  * Created by weefbellington on 10/2/14.
  */
-@Layout(R.layout.stacked_scene)
-public class StackedScene extends IndexedScene {
+
+@Layout(R.layout.new_backstack_scene_1)
+public class WelcomeScene2 extends IndexedScene {
 
     private final CrossfadeTransformer transformer;
 
-    public StackedScene() {
-        super(StackedScene.class.getName());
-        this.transformer = new CrossfadeTransformer(SampleApplication.getInstance());
+    public WelcomeScene2(Application application) {
+        super(WelcomeScene2.class.getName());
+        this.transformer = new CrossfadeTransformer(application);
+    }
+
+    public WelcomeScene2() {
+        this(SampleApplication.getInstance());
     }
 
     @Override